모듈화 진행중, inningLimit관련 버그 픽스
