+++ index.js
@@ -1,8 +1,5 @@
-const readline = require('readline');
-const rl = readline.createInterface({
-    input:process.stdin,
-    output:process.stdout
-});
+const { recordGameResult, getMinInning, getMaxInning, getModeInningLimit, getMaxInningLimit, getMinInningLimit } = require('./gameRecords');
+const {requestInput, closeInput} = require('./inputHandler');
 
 const NUMBER_LENGTH = 3;
 const GAMESTATE = {
@@ -13,24 +10,18 @@ const GAMESTATE = {
 };
 Object.freeze(GAMESTATE);
 
-//유저입력 받는 함수 -> 입력대기
-async function requestInput(msg){
-    return new Promise((resolve)=>{
-        rl.question(msg,(answer)=>resolve(answer));
-    });
-}
-
 async function playGame(){
     let gameState;
     let gameRecords = [];
-    while(gameState !== GAMESTATE.EXIT){
+    while (gameState !== GAMESTATE.EXIT){
         gameState = await requestInput('게임을 새로 시작하려면 1, 종료하려면 9를 입력하세요. ');
         gameState = parseInt(gameState);
         
-        gameState = await handleGameState(gameState, gameRecords);
+        await handleGameState(gameState, gameRecords);
     }
 }
 
+// GameState별 분기
 async function handleGameState(gameState, gameRecords) {
     switch(gameState){
         case GAMESTATE.START: 
@@ -45,14 +36,12 @@ async function handleGameState(gameState, gameRecords) {
             break;
         case GAMESTATE.EXIT:
             console.log('애플리케이션이 종료되었습니다.');
-            rl.close();
+            closeInput();
             break;
         default:
             console.log('1,2,3,9중에 입력해주세요.');
             break;
     }
-
-    return gameState;
 }
 
 function convertTime(time){
@@ -60,26 +49,30 @@ function convertTime(time){
     return `${time.getFullYear()}. ${time.getMonth()}. ${time.getDate()} ${time.getHours()}:${minutes}`;
 }
 
+// 게임 기록을 보여주는 함수
 function showHistory(gameRecords){ // 기록
     const history = gameRecords.reduce((acc, record, index) => {
         return acc + `- [${index+1}] / 시작시간: ${convertTime(record.startTime)} / 종료시간: ${convertTime(record.endTime)} / 횟수: ${record.inningLimit} / 승리자: ${record.isClear}\n`;
     }, '');
     console.log(history);
+    console.log('-------기록 종료-------');
 }
 
+// 게임 통계를 보여주는 함수
 function showStats(gameRecords){ // 통계
-  const {minInning, minInningGameIds} = getMinInning(gameRecords);
-  const {maxInning, maxInningGameIds} = getMaxInning(gameRecords);
-  const {resultModeInningLimit, resultGameIds} = getModeInningLimit(gameRecords);
-  const {minInningLimit, minInningLimitGameIds} = getMinInningLimit(gameRecords);
-  const {maxInningLimit, maxInningLimitGameIds} = getMaxInningLimit(gameRecords);
+    const {minInning, minInningGameIds} = getMinInning(gameRecords);
+    const {maxInning, maxInningGameIds} = getMaxInning(gameRecords);
+    const {resultModeInningLimit, resultGameIds} = getModeInningLimit(gameRecords);
+    const {minInningLimit, minInningLimitGameIds} = getMinInningLimit(gameRecords);
+    const {maxInningLimit, maxInningLimitGameIds} = getMaxInningLimit(gameRecords);
 
-  console.log(`1. 가장 적은 횟수: ${minInning}회 - [${minInningGameIds}]`);
-  console.log(`2. 가장 많은 횟수: ${maxInning}회 - [${maxInningGameIds}]`);
-  console.log(`3. 가장 많이 적용된 승리/패패 횟수: ${resultModeInningLimit}회 - [${resultGameIds}]`);
-  console.log(`4. 가장 큰 값으로 적용된 승리/패패 횟수: ${maxInningLimit}회 - [${maxInningLimitGameIds}]`);
-  console.log(`5. 가장 적은 값으로 적용된 승리/패패 횟수: ${minInningLimit}회 - [${minInningLimitGameIds}]`);
+    console.log(`1. 가장 적은 횟수: ${minInning}회 - [${minInningGameIds}]`);
+    console.log(`2. 가장 많은 횟수: ${maxInning}회 - [${maxInningGameIds}]`);
+    console.log(`3. 가장 많이 적용된 승리/패패 횟수: ${resultModeInningLimit}회 - [${resultGameIds}]`);
+    console.log(`4. 가장 큰 값으로 적용된 승리/패패 횟수: ${maxInningLimit}회 - [${maxInningLimitGameIds}]`);
+    console.log(`5. 가장 적은 값으로 적용된 승리/패패 횟수: ${minInningLimit}회 - [${minInningLimitGameIds}]`);
 
+    console.log('-------통계 종료-------');
 // 6. 적용된 승리/패패 횟수 평균: avg(입력제한)회
 // 7. 컴퓨터가 가장 많이 승리한 승리/패패 횟수: 컴퓨터.승리수 회
 // 8. 사용자가 가장 많이 승리한 승리/패패 횟수: 사용자.승리수 회
@@ -93,126 +86,15 @@ function showStats(gameRecords){ // 통계
 // 사용자는 -> 라운드 수
 }
 
-function getMinInning(gameRecords){
-    const minInning = gameRecords.reduce((acc, record) => {
-        if (record.inning <= acc){
-            return record.inning;
-        } else {
-            return acc;
-        }
-    }, Infinity);
-
-    let minInningGameIds = gameRecords.reduce((acc, record, index) => {
-        if (record.inning === minInning){
-            acc.push(index + 1);
-        }
-
-        return acc;
-    }, []);
-
-    return {minInning, minInningGameIds};
-}
-
-function getMaxInning(gameRecords){
-    const maxInning = gameRecords.reduce((acc, record) => {
-        if (record.inning >= acc){
-            return record.inning;
-        } else {
-            return acc;
-        }
-    }, 0);
-
-    let maxInningGameIds = gameRecords.reduce((acc, record, index) => {
-        if (record.inning === maxInning){
-            acc.push(index + 1);
-        }
-
-        return acc;
-    }, []);
-
-    return {maxInning, maxInningGameIds};
-}
-
-function getModeInningLimit(gameRecords){
-    const inningLimitFrequencies = gameRecords.reduce((acc, record, index) => {
-        if (!acc[record.inningLimit]){
-            acc[record.inningLimit] = {frequency: 0, gameIds: []};
-        }
-
-        acc[record.inningLimit].frequency += 1;
-        acc[record.inningLimit].gameIds.push(index + 1);
-        return acc;
-    }, {});
-
-    const modeInningLimit = Object.entries(inningLimitFrequencies).reduce((acc, [inningLimit, frequencyData]) =>{
-        if (frequencyData.frequency >= acc.frequencyData.frequency){
-            return {inningLimit: inningLimit, frequencyData: frequencyData};
-        }
-
-        return acc;
-    }, {inningLimit: 0, frequencyData: {frequency: 0, gameIds: []}});
-
-    const resultModeInningLimit = modeInningLimit.inningLimit;
-    const resultGameIds = modeInningLimit.frequencyData.gameIds;
-    return {resultModeInningLimit, resultGameIds};
-}
-
-function getMaxInningLimit(gameRecords){
-    const maxInningLimit = gameRecords.reduce((acc, record) => {
-        if (record.inningLimit >= acc){
-            return record.inningLimit;
-        } else {
-            return acc;
-        }
-    }, 0);
-
-    let maxInningLimitGameIds = gameRecords.reduce((acc, record, index) => {
-        if (record.inningLimit === maxInningLimit){
-            return acc.push(index + 1);
-        }
-        
-        return acc;    
-    }, []);
-
-    return {maxInningLimit, maxInningLimitGameIds};
-}
-
-function getMinInningLimit(gameRecords){
-    const minInningLimit = gameRecords.reduce((acc, record) => {
-        if (record.inningLimit <= acc){
-            return record.inningLimit;
-        } else {
-            return acc;
-        }
-    }, Infinity);
-
-    let minInningLimitGameIds = gameRecords.reduce((acc, record, index) => {
-        if (record.inningLimit === minInningLimit){
-            return acc.push(index + 1);
-        } 
-        
-        return acc;
-    }, []);
-
-    return {minInningLimit, minInningLimitGameIds};
-}
-
-function recordGameResult(gameRecords, startTime, inning, inningLimit, isClear){
-    const endTime = new Date();
-    gameRecords.push({
-        startTime,
-        endTime,
-        inning,
-        inningLimit,
-        isClear
-    });
-}
-
 async function playLoop(gameRecords) {
     console.log('컴퓨터에게 승리하기 위해 몇번만에 성공해야 하나요?');
-    const inningLimit = await requestInput("");
+    let inningLimit = await requestInput("");
+    inningLimit = parseInt(inningLimit);
+    // TODO: 0보다 큰 숫자인지 체크
+    
     const computerNumber = getComputerNumbers();
     console.log(computerNumber); // 난중에 지울것 답지임
+
     const startTime = new Date();
     
     let inning = 0;
@@ -261,38 +143,45 @@ function checkUserNumbers(userNumbers){
         console.log('숫자를 입력해주세요.');
         return null;
     }
+
     if(userNumbers.includes('0')){
         console.log('1에서 9까지의 숫자를 입력해주세요');
         return null;
     }
+
     let distinctNumbers = new Set(userNumbers);
     if (userNumbers.length !== NUMBER_LENGTH || distinctNumbers.size !== NUMBER_LENGTH){
         console.log('중복없는 세자리 숫자를 입력해주세요.');
         return null;
     }
+
     return userNumbers; 
 }
 
+// 컴퓨터 숫자와 사용자 숫자를 비교하고 strike, ball 카운트롤 반환
 function compareNumber(computerNumbers,userNumbers){
     const strikes = computerNumbers.filter((element, index)=>userNumbers.indexOf(element)===index);
     const inclusion = computerNumbers.filter((element)=>userNumbers.includes(element));
+
     const strikeCount = strikes.length;
     const ballCount = inclusion.length - strikeCount;
+
     return {strikeCount, ballCount};
 }
 
+// 게임을 클리어했는지 체크 (클리어 조건: 3 strikes)
 function checkGameClear(computerNumbers,userNumbers){
+    const {strikeCount, ballCount} = compareNumber(computerNumbers,userNumbers);
 
-    const {strikeCount, ballCount} = compareNumber(computerNumbers,userNumbers)
     if (strikeCount === NUMBER_LENGTH){
         console.log(strikeCount + '스트라이크');
         console.log('3개의 숫자를 모두 맞히셨습니다.');
         return true;
     }
+
     if (ballCount === 0 && strikeCount === 0){
         console.log('낫싱');
-    }
-    else{
+    } else { 
         console.log( ballCount + '볼 ' + strikeCount + '스트라이크');
     }
     return false;
